dynamic collision

+++ src/DynamicActor/Loader/ObjLoader.hpp
@@ -7,7 +7,7 @@ namespace dynamicactor::loader {
 
 class ObjLoader {
 public:
-    static bool loadMesh(const char* objFilepath, al::DynamicDrawActor* dynamicDrawActor, al::DynamicCollisionActor* dynamicCollisionActor, uint16_t dynamicCollisionAttributeIndex);
+    static bool loadMesh(const char* objFilepath, al::DynamicDrawActor* dynamicDrawActor, al::DynamicCollisionActor* dynamicCollisionActor, int dynamicCollisionAttributeIndex);
 };
 
 }
